Provide executor to slack_bolt app, don't re-submit
After some digging into the slack_bolt code it became evident that it's
already submitting everything into a ThreadPoolExecutor so we don't need
to do that ourselves. This reworks things to provide a custom executor
with our names and size rather than let it create its own. It also
adapts cron to handle the change. Much cleaner